backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
openai-whisper
librosa
soundfile
numpy
scikit-learn
transformers
sentence-transformers
